Move custom token: HeroCenter
