+++ packages/bodiless-demo/src/shadow/@bodiless/vital-card/Card.ts
@@ -2,15 +2,6 @@ import { H2, replaceWith } from '@bodiless/fclasses';
 import { asCardToken, vitalCardBase } from '@bodiless/vital-card';
 import { vitalTypography } from '@bodiless/vital-elements';
 
-const HeroCenter = asCardToken({
-  Theme: {
-    Wrapper: 'mt-8',
-    CTAWrapper: 'mx-auto max-w-xs',
-    TitleWrapper: 'text-center',
-    DescriptionWrapper: 'text-center',
-  },
-});
-
 const Basic = asCardToken(vitalCardBase.Basic, {
   Components: {
     TitleWrapper: replaceWith(H2),
@@ -32,5 +23,4 @@ export default {
   ...vitalCardBase,
   Hero,
   Basic,
-  HeroCenter,
 };
